+++ src/mediarithmics/index.ts
@@ -9,6 +9,7 @@ export * from './api/core/plugin/PluginPropertyInterface';
 export * from './api/core/plugin/ValueInterface';
 export * from './api/core/compartment/Compartment';
 export * from './api/core/error/MicsApiError';
+export * from './api/core/error/TransportError';
 export * from './api/datamart';
 export * from './api/plugin/audiencefeedconnector/AudienceFeedConnectorPluginResponseInterface';
 export * from './api/plugin/audiencefeedconnector/AudienceFeedConnectorRequestInterface';